% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_legend.R
\name{align_legend}
\alias{align_legend}
\title{Change alignment of ggplot legend title}
\usage{
align_legend(p, hjust = 0.5)
}
\arguments{
\item{p}{ggplot object}

\item{hjust}{Amount to adjust alignment.}
}
\description{
From Clause Wilke answer here:
https://stackoverflow.com/questions/48000292/center-align-legend-title-and-legend-keys-in-ggplot2-for-long-legend-titles
}
