% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Appendix1_var01}
\alias{Appendix1_var01}
\title{Simulated scenario where sensitivity and exposure are independent
and inter-annual temperature variation is low}
\description{
Used to estimate co-variance between sensitivity and exposure
under different conditions of inter-annual temperature variation
}
