% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ben_comment.R
\name{Ben_sim}
\alias{Ben_sim}
\title{Simulation to address Ben Sheldon's comments}
\usage{
Ben_sim(
  interpop_var,
  avg_sensitivity,
  sd_sensitivity,
  resid_sense,
  avg_exposure,
  sd_exposure,
  resid_exp,
  return = "plot"
)
}
\arguments{
\item{interpop_var}{The inter-population variation in the error of temperature
(basically causes populations to vary in the interannual variability in temp, as Ben suggested)}

\item{avg_sensitivity}{The beta of sensitivity (laying date ~ temperature)}

\item{sd_sensitivity}{The sd of the random slope term for laying date ~ temperature across pops}

\item{resid_sense}{The residual error in sensitivity (laying date ~ temperature)}

\item{avg_exposure}{The beta of exposure (temperature ~ year)}

\item{sd_exposure}{The sd of the random slope term for exposure (temperature ~ year) across pops}

\item{resid_exp}{The residual error in exposure (temperature ~ year)}

\item{return}{Should the function return a plot ("plot") or the Pearson's correlation
of sensitivity and exposure ("cor").}
}
\value{
Either a ggplot object ('plot') or numeric value representing the Pearson's correlation
between sensitivity and exposure.
}
\description{
This function will create one possible simulated relationship between sensitivity and exposure
}
