% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ben_comment.R
\name{Ben_sim_i}
\alias{Ben_sim_i}
\title{Run multiple iterations to simulate sensitivty/exposure relationship}
\usage{
Ben_sim_i(
  interpop_var = 2,
  i,
  avg_sensitivity,
  sd_sensitivity,
  resid_sense,
  avg_exposure,
  sd_exposure,
  resid_exp
)
}
\arguments{
\item{interpop_var}{The inter-population variation in the error of temperature
(basically causes populations to vary in the interannual variability in temp, as Ben suggested)}

\item{i}{Number of iterations to conduct}

\item{avg_sensitivity}{The beta of sensitivity (laying date ~ temperature)}

\item{sd_sensitivity}{The sd of the random slope term for laying date ~ temperature across pops}

\item{resid_sense}{The residual error in sensitivity (laying date ~ temperature)}

\item{avg_exposure}{The beta of exposure (temperature ~ year)}

\item{sd_exposure}{The sd of the random slope term for exposure (temperature ~ year) across pops}

\item{resid_exp}{The residual error in exposure (temperature ~ year)}
}
\description{
Run multiple iterations to simulate sensitivty/exposure relationship
}
