% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{calculate_grid_distance}
\alias{calculate_grid_distance}
\title{Calculate distance between study sites and temperature grid cells.}
\usage{
calculate_grid_distance()
}
\value{
A dataframe with distances between actual point and point used for temperature measurement.
}
\description{
For four populations (Sagunto, Spain; Cardiff, UK; Barcelona, Spain; Askainen, Sweden),
ECA&D grid data is not available at the exact study site. Instead, we used the nearest grid cell
available. This function returns information on how far these new points were from the original site.
}
\examples{
calculate_grid_distance()
}
