% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{climate_sensitivity_and_exposure}
\alias{climate_sensitivity_and_exposure}
\title{Output containing climwin slopes, pvalues and population characteristics.}
\description{
For information on how the data are generated see Prepare_data.Rmd
}
