% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{confint_2levels}
\alias{confint_2levels}
\title{Calculate 95 and 99.5 CIs for a given model.}
\usage{
confint_2levels(mod, method, model_type = "lmer", ...)
}
\arguments{
\item{mod}{A model object. Fitted with either spaMM or lmer.}

\item{method}{Character string. For lmer, should CIs be
calculated with Wald or bootstrap?}

\item{model_type}{Character string. Is the model "lmer" or "spaMM".}

\item{...}{Additional arguments passed to `confint` function
(e.g. iterations for bootstrap)}
}
\value{
A data frame with CIs for coefficients.
}
\description{
Determine 95 and 99.5 CIS for both spaMM and lmer models.
}
\examples{
#Fit test models
mod_lmer  <- lme4::lmer(disp ~ mpg + (1|cyl), data = mtcars)

#Calculate CIs
confint_2levels(mod_lmer, method = "boot", model_type = "lmer", n = 100)
}
