% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{ECAD_locations}
\alias{ECAD_locations}
\title{Lat/Long coordinates of study sites used to extract from gridded data}
\format{
A data frame with 40 rows and 4 columns
\describe{
  \item{Pop_ID}{Location of study. Site may have data on both GT and BT.}
  \item{Latitude}{Latitude of population in decimal degrees.}
  \item{Longitude}{Longitude of population in decimal degrees.}
  \item{Adjusted}{When TRUE, lat/long are adjusted slightly to overlap with
  gridded data.}
  }
}
\description{
Lat/Long coordinates of study sites used to extract from gridded data
}
