% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sensitivity.R
\name{estimate_sensitivity}
\alias{estimate_sensitivity}
\title{Estimate climatic sensitivity for all data.}
\usage{
estimate_sensitivity(
  input_data,
  first_clutch_method = "calc",
  temp_data,
  randwin = FALSE,
  repeats,
  SEM = FALSE,
  i = 1000,
  dummy_test = FALSE
)
}
\arguments{
\item{input_data}{Dataframe. Full dataset containing brood information of all birds from all populations.}

\item{first_clutch_method}{Logical. Whether first clutch is determined by data owner or >30 day rule}

\item{temp_data}{A stacked raster. Mean daily temperature data. Stored as 'tg_0.25deg_reg_v17.0' in extdata.}

\item{randwin}{Logical. Should randwin be run with climwin?}

\item{repeats}{Numeric. If randwin is TRUE, how many repeats should be run.}

\item{SEM}{Logical. Should SEM analysis also be run.}

\item{i}{Numeric. If SEM is TRUE, the number of bootstrap iterations.}

\item{dummy_test}{Logical. If TRUE, will only use the first two rows of data. Used to test code.}
}
\value{
A list with climwin and (if requested) SEM results.
}
\description{
For every species/population combination this function will:
- Format the laying date data using \code{\link{format_data}}.
- Use this format data to run climwin and detect the 'best window' where
temperature affects laying date. Using the function \code{\link{run_climwin}}
- Run randwin to determine the distribution of deltaAICc values that would
be expected from the same population with no climate signal.
Using the function \code{\link{run_climwin}}
- Run an SEM to detrend data.
Using the function \code{\link{run_SEM}}
}
\details{
Important decisions that were made as part of the process:
- We only include years where there were at least 2 values of laying date recorded.
- We only include populations where there were at least 9 years of data recorded
that met this criteria.
}
\examples{
\dontrun{
### EXAMPLE FOR HOGE VELUWE ###

#Load mean temperature data
#Temperature data file can be requested from https://www.ecad.eu/ archive
temp_data  <- raster::stack(here::here("data/unshared_files/tg_0.25deg_reg_v17.0.nc"))

#Load Brood_data.csv
input_data <- read.csv(here::here("data/unshared_files/Brood_info.csv"),
stringsAsFactors = FALSE, header = T, sep = ",")

#Create a subset of data for just Hoge Veluwe
input_data_test <- dplyr::filter(input_data, Pop_ID \%in\% "HOG")

#Run climwin and randwin
HOG_output <- estimate_sensitivity(input_data = input_data_test,
temp_data = temp_data,
randwin = TRUE, repeats = 100)
}
}
