% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Euro_temp}
\alias{Euro_temp}
\title{Raster stack of European temperatures}
\format{
A raster stack containing 24,715 individual raster levels
}
\description{
ECA&D maximum daily temperature raster stack (v16.0)
Data ranges from Jan 1st 1950 - Aug 31st 2017
}
