% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_temp_data.R
\name{extract_temp_data}
\alias{extract_temp_data}
\title{Extract temperature data for given population.}
\usage{
extract_temp_data(Lay_date_sum)
}
\arguments{
\item{Lay_date_sum}{Data frame. Summarised lay date data. This is needed to know what
date range to extract temps over. This function is generally used inside
\code{\link{run_climwin}}.}
}
\value{
R data frame object with daily mean temperature.
}
\description{
Extract temperature data for every year where laying date data are available.
}
\details{
For all populations except Vlieland and Sicily, we extract daily mean temperature
data from the ECA&D Gridded Temperature layer of Europe (v17.0).

In four populations (Sagunto, Spain; Cardiff, UK; Barcelona, Spain; Askainen, Sweden)
data were not directly available at the study site from ECA&D. Instead, we used the
nearest grid square, which was never more than 8km from the actual site. See \code{\link{calculate_grid_distance}}
for calculation.

For Vlieland, we used mean daily temperature from ROYAL NETHERLANDS METEOROLOGICAL INSTITUTE (KNMI).
From 1954, temperature data was collected on Terschelling (the neighbouring island).
From 1995, temperature data was collected directly on Vlieland.

For Sicily, we used mean temperature data collected by the data owners at the site.
}
