% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Format raw nest data into the correct format for climwin analysis.}
\usage{
format_data(input_data)
}
\arguments{
\item{input_data}{Dataframe. Raw Brood_data from formatting populations}
}
\value{
A tibble with 7 columns
}
\description{
Identify first clutches for our focal species.
}
\details{
For all population/species data:
\itemize{
\item Filter only our two focal species (GT and BT).

\item Filter only records where laying date was known.

\item Determine the cutoff (30 days after the first laying date of the season)
at which nests would no longer be considered 'first' nests. This cutoff is unique
to each species in each population in each year.

\item Classify nests as 'first' using this cutoff.

}
}
\examples{
\dontrun{
#Load Brood_info.csv
input_data <- read.csv(here::here("data/unshared_files/Brood_info.csv"),
header = T, sep = ",", stringsAsFactors = FALSE)

#Format data into correct format
format_data(input_data = input_data)
}
}
