% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_violin.R
\name{geom_flat_violin}
\alias{geom_flat_violin}
\title{Create one sided violin plot}
\usage{
geom_flat_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "dodge",
  trim = TRUE,
  scale = "area",
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{x}

\item{data}{x}

\item{stat}{x}

\item{position}{x}

\item{trim}{x}

\item{scale}{x}

\item{show.legend}{x}

\item{inherit.aes}{x}

\item{...}{x}
}
\value{
A ggproto object used in ggplot2
}
\description{
Create one sided violin plot
}
