% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Lay_date_summary}
\alias{Lay_date_summary}
\title{Primary data used as the basis for all analysis}
\format{
A data frame with 2,010 rows and 10 columns
\describe{
  \item{Pop_ID}{Unique population ID}
  \item{Species}{Species ID. Can be either GT (great tit/Parus major) or BT (blue tit/Cyanistes caeruleus)}
  \item{Sample_year}{Calendar year in which laying date was recorded.}
  \item{Mean}{Mean laying date (April days).}
  \item{SE}{Standard error of mean laying date (April days).}
  \item{Min}{Minimum observed laying date (April days).}
  \item{Count}{Total number of first nests laying date records for this population/species/year combination.}
  \item{Date}{Starting date used sliding window analysis (dd/mm/yyyy).}
  \item{SE_calc}{Inverse of standard error (1/SE), used to weight statistical models.}
  \item{Temp}{Mean temperature (oC) within population specific climate window}
  }
}
\description{
Note, in Vlieland and Sicily there are years where phenological data is
available but full temperature data is not. In these cases, Temp is NA.
}
