% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Metzger_PCA}
\alias{Metzger_PCA}
\title{Results of principal component analysis by Metzger et al. 2005 for each site.}
\format{
A data frame with 40 rows and 2 columns
\describe{
  {Pop_ID}{Unique population ID}
  {PC3_Precip}{Third principal component (precipitation)}
  }
}
\description{
Results of principal component analysis by Metzger et al. 2005 for each site.
}
