% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{MODIS_data}
\alias{MODIS_data}
\title{MODIS Terra+Aqua Combined Land Cover product (MCD12Q1) for each location in each year.}
\description{
Used to test the viability of remote sensing as an alternative method
to define habitat type. See `Step5_ReviewerComments.Rmd`.
}
