% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{MODIS_translation}
\alias{MODIS_translation}
\title{Data frame to convert MODIS number values into text categories.}
\description{
This is required to convert MODIS data into our DEC/MIX/EVE categories.
}
