% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_violin.R
\name{or_operator}
\alias{or_operator}
\alias{\%||\%}
\title{Operator to apply ifelse function}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First choice.}

\item{b}{Second choice, used if `a` is null}
}
\value{
Object a or b.
}
\description{
Operator to apply ifelse function
}
\examples{

#Return 9
9 \%||\% 10

#Return 10
NULL \%||\% 10
}
