% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Pop_info}
\alias{Pop_info}
\title{A SpatialPointsDataFrame with location information of all populations.}
\format{
A SpatialPointsDataFrame with 43 rows and 14 variables
\describe{
  \item{Study_ID}{Three letter code for study site. In some cases, multiple populations are administered in the same study.}
  \item{Pop_ID}{Three letter code specifying the population ID.}
  \item{Site_Name}{Full name of population.}
  \item{Latitude}{Latitude of population in decimal degrees.}
  \item{Latitude_min}{Latitude of population minutes.}
  \item{Longitude}{Longitude of population in decimal degrees.}
  \item{Longitude_min}{Longitude of population minutes.}
  \item{Sunrise}{Time of sunrise (hh:mm:ss) on April 1st from https://www.esrl.noaa.gov/gmd/grad/solcalc/sunrise.html.}
  \item{Sunset}{Time of sunset (hh:mm:ss) on April 1st from https://www.esrl.noaa.gov/gmd/grad/solcalc/sunrise.html.}
  \item{Number_boxes}{Number of nest boxes used in the population (if provided).}
  \item{Area}{Area of study site in m2.}
  \item{Habitat_Type}{Deciduous (DEC), Mixed (MIX), or Evergreen (EVE).}
  \item{Reference}{Citation with information on the study site, if applicable.}
  \item{Habitat_Notes}{Notes on habitat used to classify Habitat_Type measure.}
  }
}
\description{
Site location for all populations in decimal degrees including site information.
}
