% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Population_summary}
\alias{Population_summary}
\title{A mean laying date for every species/population combo}
\format{
A tibble with 67 rows and 4 variables
\describe{
  \item{Pop_ID}{Three letter code specifying the population ID.}
  \item{Species}{Great tit (GT) or blue tit (BT).}
  \item{mean_LD}{Mean laying date (April days).}
  \item{mean_LD_Julian}{Mean laying date (Julian days)}
  }
}
\description{
A summary table showing mean laying date in April days and Julian days for
every species/population combination.
}
