% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_climwin.R
\name{run_climwin}
\alias{run_climwin}
\title{Run climwin analysis on data for a single population/species combo.}
\usage{
run_climwin(
  input_data,
  temp_data,
  randwin = FALSE,
  repeats = NA,
  dummy_test = FALSE
)
}
\arguments{
\item{input_data}{A data frame with laying date data for a given population and species.}

\item{temp_data}{Raster stack of temperature data across Europe.}

\item{randwin}{Logical (TRUE/FALSE). Should randwin be run as well as slidingwin?}

\item{repeats}{If randwin is TRUE, the number of times that data will be randomised.}

\item{dummy_test}{To test code. If TRUE, will only analyse the first two years of data. Used for testing code.}
}
\value{
A list containing the output of climwin and (if conducted) randwin.
}
\description{
Run slidingwin and (optionally) randwin on data from a given bird population.
This is generally used inside function \code{\link{estimate_sensitivity}}.
}
\details{
As part of this approach we:
\itemize{

\item Calculate the annual mean laying date for the focal population.

\item Calculate standard error of the laying date mean that will
be used for weighting in models.

\item Exclude years with only 1 nest (where no SE could be calculated).

\item Extract mean daily temperature data using \code{\link{extract_temp_data}}.

\item For Vlieland and Sicily, remove records where climatic data is unavailable.

\item Run climwin with a baseline model: laying date ~ year, weight = 1/SE.

\item Our climate windows test for a linear relationship between mean temperature
and annual mean laying date up to 365 days before June 1st. Any missing climatic data
is interpolated by taking the mean value of temperature on that date in all other years.

\item Run randwin using the same setup as climwin.

}
}
