% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SEM.R
\name{run_SEM}
\alias{run_SEM}
\title{Run SEM for a given dataset.}
\usage{
run_SEM(
  Clim,
  Lay_date_sum,
  dataset,
  dummy_test = FALSE,
  i = 1000,
  internal = TRUE
)
}
\arguments{
\item{Clim}{Vector. Climate data for this population}

\item{Lay_date_sum}{Vector. Lay date data for this population (average first laydate for each year)}

\item{dataset}{Data frame. Dataframe with best temperature windows for each population/species}

\item{dummy_test}{Logical. If TRUE, will not save the RDS file.}

\item{i}{Numeric. Number of bootstrap iterations for the SEM model.}

\item{internal}{Logical. Is the function being run internally or called directly?}
}
\value{
Output a model summary and save an RDS file with SEM results
}
\description{
Run a structural equation model for a given population/species combination
to disentangle the effect of temperature on laying date from
potential shared trends in the data.
}
\details{
This process will:
\itemize{

\item Use the function singlewin to return to data from the best climate window
for a given population. Use the same structure as the climwin analysis (see \code{\link{run_climwin}}).

\item Create an SEM model with multiple pathways that accounts for a) Temp changing over time,
b) Laying date changing over time, c) Temp affecting laying date.

\item Bootstrap with i iterations to estimate coefficients.

}
}
