% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Sicily_temp_data}
\alias{Sicily_temp_data}
\title{Data collected on site in the Sicily population from 01/01/2006 - 31/12/2017.}
\format{
A data frame with 4,383 rows and 4 columns
\describe{
 {Year}{Year of temperature data collection}
 {Month}{Month of temperature data collection}
 {Day}{Day of temperature data collection}
 {Temperature}{Mean daily temperature (oC)}
  }
}
\description{
Data collected on site in the Sicily population from 01/01/2006 - 31/12/2017.
}
