% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_ubuntu}
\alias{theme_ubuntu}
\title{Personal theme for making plots}
\usage{
theme_ubuntu(
  base_size = 12,
  base_family = "Ubuntu",
  base_line_size = base_size/20,
  base_rect_size = base_size/20,
  legend = "none"
)
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}

\item{base_line_size}{Standard line size}

\item{base_rect_size}{Standard rectangle size}

\item{legend}{Default value for theme(legend.position)}
}
\description{
Personal theme for making plots
}
\examples{
library(ggplot2)

#Don't use Ubuntu font because it may not be loaded
ggplot() +
geom_point(data = mtcars, aes(x = disp, y = mpg)) +
theme_ubuntu(base_family = "sans")
}
