% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Vlieland_temp_data_new}
\alias{Vlieland_temp_data_new}
\title{Data collected on Vlieland by KNMI from 01/01/1996 - 26/07/2017.}
\format{
A data frame with 7,828 rows and 5 columns
\describe{
  {Station_no}{KNMI station number}
  {Date}{Date of weather data (yyyymmdd)}
  {Mean}{Daily mean temperature (oC)}
  {Min}{Daily minimum temperature (oC)}
  {Max}{Daily maximum temperature (oC)}
  }
}
\description{
Data collected on Vlieland by KNMI from 01/01/1996 - 26/07/2017.
}
