% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Vlieland_temp_data_old}
\alias{Vlieland_temp_data_old}
\title{Data collected on Terschelling by KNMI from 01/01/1954 - 31/12/2016.}
\format{
A data frame with 23,011 rows and 5 columns
\describe{
  {STN}{KNMI station number}
  {Date}{Date of weather data (dd/mm/yyyy)}
  {Mean}{Daily mean temperature (oC)}
  {Min}{Daily minimum temperature (oC)}
  {Max}{Daily maximum temperature (oC)}
  }
}
\description{
Data collected on Terschelling by KNMI from 01/01/1954 - 31/12/2016.
}
